Add version to aggregator API config endpoint
